Infosys-Agentic-Foundry-Backend/knowledgebase_server/requirements.txt
fastapi
uvicorn[standard]
asyncpg
numpy
python-dotenv
requests
python-multipart
PyPDF2
python-docx
PyMuPDF
Pillow
rapidocr-onnxruntime